requirements.txt
pyserial
python-vlc